+++ psycop/common/sequence_models/__init__.py
@@ -1,5 +1,6 @@
 from .dataset import PatientDataset
-from .embedders.BEHRT_embedders import BEHRTEmbedder, Embedder
+from .embedders.BEHRT_embedders import BEHRTEmbedder
+from .embedders.interface import Embedder
 from .tasks import BEHRTForMaskedLM
 
 __all__ = [
